rename ViewComponent.class to ViewWithFragmentComponent.class and ViewComponentBuilder.class to ViewWithFragmentComponentBuilder.class
RELNOTES=N/A

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=310232879
